master detail page rename
