+++ templates/Web/Pages/Angular.MasterDetail/src/app/app-shell/AngularMasterDetail/master-detail-page/master-detail-page.component.ts
