+++ templates/Web/Pages/Angular.MasterDetail/src/app/app-shell/AngularMasterDetail/master-detail.service.ts
@@ -2,7 +2,7 @@
 import { CONSTANTS } from 'src/constants';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { IMasterDetailText } from './MasterDetailPage/master-detail-page.component';
+import { IMasterDetailText } from './master-detail-page/master-detail-page.component';
 
 @Injectable({
   providedIn: 'root'
